+++ k6/utils/constants.ts
@@ -1,0 +1,4 @@
+/**
+ * Telegram webhook url for API calls
+ */
+export const K6_WEBHOOK_URL = "http://host.docker.internal:3001/chat-adviser-bot";