+++ prisma/migrations/20240212123612_1_17_0/migration.sql
@@ -1,0 +1,5 @@
+-- CreateIndex
+CREATE INDEX "chats_type_idx" ON "chats"("type");
+
+-- CreateIndex
+CREATE INDEX "votebans_is_completed_idx" ON "votebans"("is_completed");