+++ src/messages/messages.service.spec.ts
@@ -1,7 +1,9 @@
+import { CacheModule } from "@nestjs/cache-manager";
 import { Test } from "@nestjs/testing";
 import { TelegrafModule } from "nestjs-telegraf";
 import { BOT_TOKEN } from "src/app.constants";
 import { PrismaModule } from "src/prisma/prisma.module";
+import { cache } from "src/utils/cache";
 
 import { MessagesService } from "./messages.service";
 
@@ -10,7 +12,11 @@ describe("MessagesService", () => {
 
   beforeEach(async () => {
     const testingModule = await Test.createTestingModule({
-      imports: [TelegrafModule.forRoot({ launchOptions: false, token: BOT_TOKEN }), PrismaModule],
+      imports: [
+        CacheModule.register({ isGlobal: true, store: cache }),
+        TelegrafModule.forRoot({ launchOptions: false, token: BOT_TOKEN }),
+        PrismaModule,
+      ],
       providers: [MessagesService],
     }).compile();