+++ src/voteban/voteban.service.spec.ts
@@ -1,8 +1,10 @@
+import { CacheModule } from "@nestjs/cache-manager";
 import { Test } from "@nestjs/testing";
 import { TelegrafModule } from "nestjs-telegraf";
 import { BOT_TOKEN } from "src/app.constants";
 import { PrismaModule } from "src/prisma/prisma.module";
 import { SettingsModule } from "src/settings/settings.module";
+import { cache } from "src/utils/cache";
 
 import { VotebanService } from "./voteban.service";
 
@@ -11,7 +13,12 @@ describe("VotebanService", () => {
 
   beforeEach(async () => {
     const testingModule = await Test.createTestingModule({
-      imports: [TelegrafModule.forRoot({ launchOptions: false, token: BOT_TOKEN }), PrismaModule, SettingsModule],
+      imports: [
+        CacheModule.register({ isGlobal: true, store: cache }),
+        TelegrafModule.forRoot({ launchOptions: false, token: BOT_TOKEN }),
+        PrismaModule,
+        SettingsModule,
+      ],
       providers: [VotebanService],
     }).compile();