+++ test/settings.e2e-spec.ts
@@ -3,7 +3,7 @@ import { Test } from "@nestjs/testing";
 import { http, HttpResponse } from "msw";
 import request from "supertest";
 import type { App } from "supertest/types";
-import { server } from "test/utils/setup-after-env";
+import { server } from "test/utils/server";
 
 import { AppModule } from "../src/app.module";
 import * as fixtures from "./fixtures/settings";
@@ -19,6 +19,23 @@ describe("SettingsModule (e2e)", () => {
     const moduleFixture = await Test.createTestingModule({ imports: [AppModule] }).compile();
     app = moduleFixture.createNestApplication();
     await app.init();
+  });
+
+  it("should not render chats as an answer to /mychats command in a supergroup chat", async () => {
+    let sendMessagePayload;
+    server.use(
+      http.post(`${TELEGRAM_API_BASE_URL}/sendMessage`, async (info) => {
+        sendMessagePayload = await info.request.json();
+        return HttpResponse.json({ ok: true });
+      }),
+    );
+
+    const response = await request(TEST_WEBHOOK_BASE_URL)
+      .post(TEST_WEBHOOK_PATH)
+      .send(fixtures.myChatsInSupergroupWebhook);
+
+    expect(response.status).toBe(200);
+    expect(sendMessagePayload).toEqual(fixtures.myChatsInSupergroupSendMessagePayload);
   });
 
   it("should prompt admin to make settings when adding the bot to a new chat", async () => {
@@ -64,6 +81,22 @@ describe("SettingsModule (e2e)", () => {
     expect(sendMessagePayload2).toEqual(fixtures.groupCreatedSendMessagePayload);
   });
 
+  it("should refresh chats", async () => {
+    let editMessageTextPayload;
+    server.use(
+      http.post(`${TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
+        editMessageTextPayload = await info.request.json();
+        return HttpResponse.json({ ok: true });
+      }),
+    );
+
+    const response = await request(TEST_WEBHOOK_BASE_URL).post(TEST_WEBHOOK_PATH).send(fixtures.cbRefreshWebhook);
+
+    expect(response.status).toBe(200);
+    expect(response.body).toEqual({ callback_query_id: "1", method: "answerCallbackQuery" });
+    expect(editMessageTextPayload).toEqual(fixtures.cbRefreshEditMessageTextPayload);
+  });
+
   it("should render chats as an answer to /mychats command in a private chat", async () => {
     let sendMessagePayload;
     server.use(
@@ -81,20 +114,19 @@ describe("SettingsModule (e2e)", () => {
     expect(sendMessagePayload).toEqual(fixtures.myChatsInPrivateChatSendMessagePayload);
   });
 
-  it("should not render chats as an answer to /mychats command in a supergroup chat", async () => {
-    let sendMessagePayload;
+  it("should render the second page of chats", async () => {
+    let editMessageTextPayload;
     server.use(
-      http.post(`${TELEGRAM_API_BASE_URL}/sendMessage`, async (info) => {
-        sendMessagePayload = await info.request.json();
+      http.post(`${TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
+        editMessageTextPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
     );
 
-    const response = await request(TEST_WEBHOOK_BASE_URL)
-      .post(TEST_WEBHOOK_PATH)
-      .send(fixtures.myChatsInSupergroupWebhook);
+    const response = await request(TEST_WEBHOOK_BASE_URL).post(TEST_WEBHOOK_PATH).send(fixtures.cbChatsWebhook);
 
     expect(response.status).toBe(200);
-    expect(sendMessagePayload).toEqual(fixtures.myChatsInSupergroupSendMessagePayload);
+    expect(response.body).toEqual({ callback_query_id: "1", method: "answerCallbackQuery" });
+    expect(editMessageTextPayload).toEqual(fixtures.cbChatsEditMessageTextPayload);
   });
 });